chore: trim onboarding-agent legacy routes and obsolete helpers
