--- app/api/onboarding-agent/sessions/[sessionId]/files/route.ts
@@ -1,20 +1,0 @@
-import { NextResponse } from "next/server";
-
-type RouteContext = {
-  params: Promise<{
-    sessionId: string;
-  }>;
-};
-
-export async function POST(_: Request, context: RouteContext) {
-  const { sessionId } = await context.params;
-
-  return NextResponse.json(
-    {
-      ok: false,
-      error: "Legacy file registration endpoint has been retired. Use /api/onboarding-agent/sessions/[sessionId]/upload-files.",
-      sessionId,
-    },
-    { status: 410 },
-  );
-}
